Added Implementation For SignUp
added a basic implementation for user sign up. This implementation checks for specific edges cases and if no error is found then it creates a user in the database and returns an access token, which that user can use to access the resources of the REST API. All the failing tests of the previous issue are now succeeding, which follows the TDD principle/flow.