Added Implementation For Log Out
Added a basic implementation for log out. This implementation first validates and decodes the token to make sure the user can log out. Once that is done all the existing tokens are put in a blacklist by incrementing by 1 the user specific counter that exists in the database(each user has 1 once registered). Because in the validation there is a check that checks that specific field all previous tokens with the previous counter number will fail if used to access any resources of the REST API after a logout process has occured. After some thought I also added an extra test that checks for this very specific case, which passes(the other 3 pass too) and thus confirms that this implementation works.