Added Failing Unit Tests For Create Todo Functionality
I added 2 unit tests of which 1 tests for the edge case that the database is down/throws exception and the other test checks the cenario in which the todo should be created successfully. In the future additional tests might be added, but for now these are all I can think of.