Added Implementation For Create Todo
added a basic implementation that allows a user to create todos. I also removed one of the tests I added in the previous commit, which would check the edge case of a database failure, because unit testing entity framework is basically impossible without the in memory nugget package, which unfortunately does not allow for exception throwing.