Added Create Todo API Endpoint
I added a create todo endpoint that allows the user after authentication to create a todo. The implementation follows the best practises, which means that a 201 status code is returned with the created object and the location of where the user can access the newly created resource. I will need to refactor this implementation a little bit after I add the GetTodoById endpoint, but for now it works well.