Added Integration Endpoints For Create Todo Endpoint
I added integration tests that check a basic flow in which a user tries to create a todo. One of the tests specificially tests that a user that is not authenticated can not create a todo. These tests should make sure that test regressions should not happen with code additions or refactoring.