Added Failing Unit Tests For Get Todos Functionality
I added 2 unit tests that test the basic flow of the GetTodos method. In general it will either return no todos, because the user has none or it will return the todos if user have some. As before I am not going to test the exception, because the in memory database makes it too difficult to do so.