Added Integration Tests For Get Todos Endpoint
I added 2 integration check for the GetTodos that check what happens if the user is not authenticated and if the user is authenticated. The tests pass, which means that everything should work as intended and these tests will make sure that in the future there will not be any code regressions.