Added Failing Unit Tests For Get Todos By Id Functionality
I added 3 basic tests that test the basic of the GetTodoById method. The first 2 tests check the scenarios in which either the todo does not exist or the user does not own it and thus the result should be null and the third test checks the success case. As mentioned before I am not going to check the exception case, because it is too difficult using the in memory database.