Added Integration Tests For Get Todo By Id Endpoint
I added 3 integration tests for the Get Todo By Id endpoint that check typical scenarios. The first 2 check what happens if the user is not authenticated and if the todo does not exist and the 3rd test checks the success case. These integration tests should make sure that future refactoring, which I will probably do soon, will not lead to code regressions.