Added Implementation For Update Todo
I added a basic implementation to update the todo in the the database(in general the Title and the IsDone field). Maybe I should make protection, so that the user can not change the userId field, because if they do they will not be able to access that Todo ever again. Other than that I added a NotFound error code that I also added to the getTodoByIdAsync method, so that it is more explicit when a todo was of the given id is not found.