Added Integrations Tests For UpdateTodo Endpoint
I added 3 integration tests for the update todo endpoint, which are pretty similar to the integrations tests of the GetTodo Endpoint. The first 2 test the error case, which occur if the access token is invalid and/or if the user does not own the todo and/or the todo does not exist. The 3rd test checks the success case. These tests will help ensure that future code regressions will not occur so easily.