Added Implementation For Delete Todo
I added a basic implementation that allows the user to delete 1 of their todos using its id. The implementation is pretty similar to the GetUserTodoByIdAsync and UpdateTodo with the major difference between that the delete does not return the deleted todo since the REST API does not need to return it to the user.