Added Integration Tests For DeleteTodo Endpoint
I added 3 integration tests for the delete todo endpoint, which are pretty similar to the integration tests of the GetTodo endpoint and of the UpdateTodo endpoint. The first 2 check for the typical error cases, which are the todo not existing or the user not owning the todo and the third test as usual checks the success case, in which the todo should be deleted. I also refactored the create todo endpoint, which now creates the location of the newly created todo dynamically(no longer localhost stuff).