Added Implementation For Create Todo Item
I added a basic implementation that allows the user to create a todo item for one of their todos. If the todo does not exist then the todo item can not be created. I also refactored the code a little bit to accomodate the new changes(added a List<TodoItems> as property in the todo) and I also restructured the files a little bit.