Added Create Todo Item API Endpoint
I added a create todo item that allows an authentication user to create a todo item for one of their todos. I will just need to refactor the location where the new resource can be found, because as of now it just returns the current localhost. Other than that the API returns appropriates status codes for different possible error cases.