Added Integration Tests For CreateTodoItem Endpoint
I added 3 integration tests for the create todo endpoint, which test the basic flow of a user who tries to create a todo item for one of their todos. The first 2 tests check typical error cases, which might occur either because the access token is invalid or either beecause the todo does not exist or the todo does not belong to the user. The third test checks the success case, in which the create todo item should be created and added to the todo.