Added Integration Tests For The GetTodoItemById API Endpoint
I added 4 integrations tests for the GetTodoItemById API endpoint that check typical scenarios. The first 2 check the error case, in which either the access token is invalid or the user does not own the todo or it does not exist. The third tests checks the error case, in which the todo item itself does not exist. Finally the fourth test checks the success case, in which the todo should be returned successfully to the user.