Added Failing Unit Tests For Update Todo Item Functionality
I added 4 unit tests that test the UpdateUserTodoItemAsync method. These unit tests are pretty similar to the unit tests of the GetUserTodoItemAsync method. The first 2 tests check the error case, in which the user does not own the todo or the todo does not exist. The third test checks the error case, in which the todo of the user exists, but the todo item itself does not exist. Finally the fourth test checks the success case, in which the todo is successfully updated. Following TDD, the tests are failing for now.