Added Implementation For Update Todo Item
I added a basic implementation in the data access library that allows the user to update one of their todo items(Title, Description and IsDone fields) that exist inside one of their todos using the todo items id. Also the failing unit tests of the previous commit are now passing.