Added Integration Tests For UpdateTodoItem Endpoint
I added 4 integrations tests for the UpdateTodoItem endpoint that check usual cases. The first 2 check the error cases, in which the access token does not exist or the todo does not exist or it exists but does not belong to the user. The third test check the error case, in which the user todo exists, but the todo item itself does not exist. Finally the fourth case is the success case, in which the todo should be updated successfully.