Added Failing Unit Tests For Delete Todo Item Functionality
I added 4 failing tests for the method DeleteUserTodoItemAsync in the DataAccess library. The first 2 check what happens for the error case, in which the todo does not exist or the user does not own it. The third test checks the scenario, in which the user todo exists, but the todo item itself does not exist. Finally the fourth case is the success case, in which the todo item needs to be deleted successfully.