Added DeleteTodoItem API Endpoint
I added a DeleteTodoItem endpoint in the rest API that can be used by a user to delete a todo item that exists inside one of their todos. Also the endpoint returns appropriate status codes for different possible scenarios. Finally I refactored the CreateTodoItem endpoint, so that now it creates the location of the newly create todo dynamically and not statically.