Added Integration Tests For DeleteTodoItem Endpoint
I added 4 integrations tests for the DeleteTodoItem endpoint. The first 2 check the error cases, in which either the access token is invalid or the todo does not exist/does not belong to the user of the access token. The third test checks the error case, in which the todo exists, but the todo item itself does not exist. Finally the fourth tests checks the success case in which the todo should be correctly deleted and NoContent status code should be returned. These integration tests will hopefully make sure that future changes will not lead to code regressions.